invalid mseed

+++ nzgmdb/scripts/real_time_eq_runs.py
@@ -237,9 +237,7 @@ def run_event(
     conda_sh: Annotated[Path, typer.Argument(exists=True, file_okay=True)],
     gmc_activate: Annotated[str, typer.Argument()],
     gmc_predict_activate: Annotated[str, typer.Argument()],
-    ko_matrix_path: Annotated[
-        Path, typer.Argument(exists=True, file_okay=False)
-    ],
+    ko_matrix_path: Annotated[Path, typer.Argument(exists=True, file_okay=False)],
     add_seismic_now: Annotated[bool, typer.Option(is_flag=True)] = False,
     machine: Annotated[
         cfg.MachineName,
@@ -550,8 +548,10 @@ def poll_earthquake_data(
                 )
 
                 if not result:
-                    # remove the event directory
-                    shutil.rmtree(event_dir)
+                    # Check if the event directory exists before trying to remove it, in case the error happened before the directory was created
+                    if event_dir.exists():
+                        # remove the event directory
+                        shutil.rmtree(event_dir)
                     # add the event to the no stations events
                     no_stations_events.append(event_id)
                 init_start_date = end_date
